feat: 문진표 검사 등록 API 구현

+++ src/main/java/com/example/pawpaw/domain/survey/dto/response/SurveyItemResponse.java
@@ -1,7 +1,7 @@
 package com.example.pawpaw.domain.survey.dto.response;
 
 public record SurveyItemResponse(
-    Long surveyId,
+    int surveyId,
     String title,
     int minAgeMonths,
     int maxAgeMonths
